Clear search highlight after clicking on a node
