+++ plugins/SBOMVis/customTemplates/src/sidebar.js
@@ -1,5 +1,6 @@
 import { zoomToView } from "#buttonEventHandlersModule";
 import { insert, isLeaf, resolvePath } from "#kvGraph";
+import { setGraphColor } from "#utilsModule";
 
 function createRow(
 	leftColumnValue,
@@ -369,20 +370,6 @@ export function insertSearchSidebar(id) {
 	fragment.appendChild(resultsSection);
 
 	rootNode.replaceChildren(fragment);
-
-	function setGraphColor(color) {
-		const nodesDataset = nodes.get({ returnType: "Object" });
-
-		for (const nID in nodesDataset) {
-			nodesDataset[nID].color = color;
-		}
-
-		const tmp = [];
-		for (const nID in nodesDataset)
-			if (Object.hasOwn(nodesDataset, nID)) tmp.push(nodesDataset[nID]);
-
-		nodes.update(tmp);
-	}
 
 	function createResultsSection(nodeIDs) {
 		function createResultsCard(nodeID) {
