+++ plugins/SBOMVis/customTemplates/src/utils.js
@@ -1,0 +1,13 @@
+export function setGraphColor(color) {
+	const nodesDataset = nodes.get({ returnType: "Object" });
+
+	for (const nID in nodesDataset) {
+		nodesDataset[nID].color = color;
+	}
+
+	const tmp = [];
+	for (const nID in nodesDataset)
+		if (Object.hasOwn(nodesDataset, nID)) tmp.push(nodesDataset[nID]);
+
+	nodes.update(tmp);
+}
